+++ i3configger/context.py
@@ -9,6 +9,13 @@ log = logging.getLogger(__name__)
 
 
 def process(item: t.Iterable[t.Union[dict, partials.Partial]]) -> dict:
+    ctx = merge_all(item)
+    ctx = resolve_variables(ctx)
+    ctx = remove_variable_markers(ctx)
+    return ctx
+
+
+def merge_all(item: t.Iterable[t.Union[dict, partials.Partial]]) -> dict:
     """merge contexts of an arbitrary number of items"""
     dicts = []
     for elem in item: